@@ -93,46 +103,45 @@
     "testEnvironment": "node"
   },
   "dependencies": {
-    "chalk": "^2.4.2",
-    "got": "^9.6.0",
-    "minimist": "^1.2.0",
-    "minimist-options": "^4.0.1",
-    "mkdirp": "^0.5.1",
-    "ora": "^3.4.0",
+    "chalk": "^4.1.0",
+    "execa": "^4.0.2",
+    "got": "^11.3.0",
+    "lodash": "^4.17.15",
+    "minimist": "^1.2.5",
+    "minimist-options": "^4.1.0",
+    "mkdirp": "^1.0.4",
+    "ora": "^4.0.4",
     "redent": "^3.0.0",
-    "rimraf": "^2.6.3",
+    "semver": "^7.3.2",
     "tildify": "^2.0.0",
-    "untildify": "^4.0.0",
-    "zce": "^0.1.2"
+    "untildify": "^4.0.0"
   },
   "devDependencies": {
-    "@commitlint/cli": "^7.6.1",
-    "@commitlint/config-conventional": "^7.6.0",
-    "@types/execa": "^0.9.0",
-    "@types/got": "^9.4.4",
-    "@types/jest": "^24.0.12",
-    "@types/lodash": "^4.14.126",
+    "@types/execa": "^2.0.0",
+    "@types/got": "^9.6.11",
+    "@types/jest": "^26.0.0",
+    "@types/lodash": "^4.14.156",
     "@types/minimist": "^1.2.0",
-    "@types/mkdirp": "^0.5.2",
-    "@types/node": "^12.0.0",
-    "@types/rimraf": "^2.0.2",
-    "@types/semver": "^6.0.0",
-    "@typescript-eslint/eslint-plugin": "^1.8.0",
-    "@typescript-eslint/parser": "^1.8.0",
-    "commitizen": "^3.1.1",
-    "cz-conventional-changelog": "^2.1.0",
-    "eslint": "^5.16.0",
-    "execa": "^1.0.0",
-    "husky": "^2.2.0",
-    "jest": "^24.8.0",
-    "lint-staged": "^8.1.6",
-    "prettier": "^1.17.0",
-    "standard-version": "^6.0.1",
-    "ts-jest": "^24.0.2",
-    "ts-node": "^8.1.0",
-    "typescript": "^3.4.5"
+    "@types/mkdirp": "^1.0.1",
+    "@types/node": "^14.0.13",
+    "@types/rimraf": "^3.0.0",
+    "@types/semver": "^7.2.0",
+    "@typescript-eslint/eslint-plugin": "^3.4.0",
+    "@typescript-eslint/parser": "^3.4.0",
+    "eslint": "^7.3.1",
+    "eslint-config-standard": "^14.1.1",
+    "eslint-plugin-import": "^2.21.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.1",
+    "jest": "^26.0.1",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.0.5",
+    "rimraf": "^3.0.2",
+    "ts-jest": "^26.1.1",
+    "typescript": "^3.9.5"
   },
   "engines": {
-    "node": ">=8.10"
+    "node": ">=10"
   }
 }
