@@ -1,10 +1,7 @@
 import { homedir, tmpdir } from 'os'
 import { join } from 'path'
-
 import execa from 'execa'
-
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const { name } = require('../../../package.json')
+import { name } from '../../../package.json'
 
 const identify = process.env.NODE_ENV === 'test' ? name : `${name}-test`
 
@@ -12,7 +9,7 @@ const identify = process.env.NODE_ENV === 'test' ? name : `${name}-test`
  * Get data path.
  * @param paths path
  */
-export const getDataPath = (...paths: string[]) => {
+export const getDataPath = (...paths: string[]): string => {
   return join(homedir(), '.config', identify, ...paths)
 }
 
@@ -20,7 +17,7 @@ export const getDataPath = (...paths: string[]) => {
  * Get temp path.
  * @param paths path
  */
-export const getTempPath = (...paths: string[]) => {
+export const getTempPath = (...paths: string[]): string => {
   return join(tmpdir(), identify, ...paths)
 }
 
