@@ -1,7 +1,5 @@
 import * as sniffer from './sniffer'
-
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const { engines } = require('../../package.json')
+import { engines } from '../../package.json'
 
 let log: jest.SpyInstance
 let exit: jest.SpyInstance
@@ -21,7 +19,7 @@ test('unit:core:sniffer', async () => {
 })
 
 test('unit:core:sniffer:sniff', async () => {
-  await sniffer.sniff()
+  sniffer.sniff()
 
   expect(log).not.toHaveBeenCalled()
 })
@@ -30,14 +28,12 @@ test('unit:core:sniffer:sniff:version', async () => {
   const originVersion = engines.node
   engines.node = '>888.8.8'
 
-  await sniffer.sniff()
+  sniffer.sniff()
 
   expect(log.mock.calls[0][0]).toBe(
     'You are using Node.js %s, but this version of %s requires Node.js %s.'
   )
-  expect(log.mock.calls[1][0]).toBe(
-    'Please upgrade your Node.js version before this operation.'
-  )
+  expect(log.mock.calls[1][0]).toBe('Please upgrade your Node.js version before this operation.')
   expect(exit.mock.calls[0][0]).toBe(1)
 
   engines.node = originVersion
