@@ -1,7 +1,6 @@
 import { runCommand } from './utils'
 
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const { name, version } = require('../package.json')
+import { name, version } from '../package.json'
 
 jest.setTimeout(8000)
 
