@@ -1,11 +1,11 @@
 import { unknownCommand } from '../error'
-
 import { Command, Context } from '../types'
+import { description } from '../../../package.json'
 
 const command: Command = {
   name: 'default',
   usage: '<command> [options]',
-  description: require('../../../package.json').description,
+  description: description,
   hidden: true,
   action: async (ctx: Context) => {
     if (!ctx.primary) return
