@@ -5,10 +5,7 @@ import { logger } from './helpers'
  * @param name command name
  * @param fallback fallback command tip
  */
-export const unknownCommand = (
-  name: string,
-  fallback: string = '[bin] --help'
-) => {
+export const unknownCommand = (name: string, fallback = '[bin] --help'): void => {
   logger.error('Unknown command: `%s`.', name)
   logger.error('Type `%s` to view all commands.', fallback)
   process.exit(1)
@@ -18,23 +15,23 @@ export const unknownCommand = (
  * Throw missing option error.
  * @param name argument name
  */
-export const missingArgument = (name: string) => {
+export const missingArgument = (name: string): void => {
   logger.error('Missing required argument: `<%s>`.', name)
   process.exit(1)
 }
 
-// export const unknownOption = (flag) => {
+// export const unknownOption = (flag: string): void => {
 //   if (this._allowUnknownOption) return
 //   console.error('Unknown option: `%s`.', chalk.yellow(flag))
 //   process.exit(1)
 // }
 
-// export const missingArgument = (name) => {
+// export const missingArgument = (name: string): void => {
 //   console.error('Missing required argument: `%s`.', chalk.yellow(`<${name}>`))
 //   process.exit(1)
 // }
 
-// export const optionMissingArgument = (option, flag) => {
+// export const optionMissingArgument = (option, flag): void => {
 //   if (flag) {
 //     console.error('Missing required argument for option: `%s`, got `%s`', chalk.yellow(option.flags), chalk.yellow(flag))
 //   } else {
@@ -43,7 +40,7 @@ export const missingArgument = (name: string) => {
 //   process.exit(1)
 // }
 
-// export const variadicArgNotLast = (name) => {
+// export const variadicArgNotLast = (name): void => {
 //   console.error('Variadic arguments must be last: `%s`.', chalk.yellow(name))
 //   process.exit(1)
 // }
