@@ -1,11 +1,8 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import { cursorTo, clearScreenDown } from 'readline'
 
 import chalk from 'chalk'
 import redent from 'redent'
 import ora, { Ora } from 'ora'
-
-import { DynamicObject } from '../types'
 
 // Disable colors output for testing
 chalk.enabled = process.env.NODE_ENV !== 'test'
@@ -14,7 +11,7 @@ chalk.enabled = process.env.NODE_ENV !== 'test'
  * Writes a normal message.
  * @param message The message to show.
  */
-export const log = (message?: any, ...optionalParams: any[]) => {
+export const log = (message?: unknown, ...optionalParams: unknown[]): void => {
   console.log(message, ...optionalParams)
 }
 
@@ -23,7 +20,7 @@ export const log = (message?: any, ...optionalParams: any[]) => {
  * This is the default type you should use.
  * @param message The message to show.
  */
-export const info = (message?: any, ...optionalParams: any[]) => {
+export const info = (message?: unknown, ...optionalParams: unknown[]): void => {
   log(chalk.reset(message), ...optionalParams)
 }
 
@@ -32,7 +29,7 @@ export const info = (message?: any, ...optionalParams: any[]) => {
  * When something is successful.  Use sparingly.
  * @param message The message to show.
  */
-export const success = (message?: any, ...optionalParams: any[]) => {
+export const success = (message?: unknown, ...optionalParams: unknown[]): void => {
   log(chalk.green(message), ...optionalParams)
 }
 
@@ -41,7 +38,7 @@ export const success = (message?: any, ...optionalParams: any[]) => {
  * This is when the user might not be getting what they're expecting.
  * @param message The message to show.
  */
-export const warn = (message?: any, ...optionalParams: any[]) => {
+export const warn = (message?: unknown, ...optionalParams: unknown[]): void => {
   log(chalk.yellow(message), ...optionalParams)
 }
 
@@ -50,7 +47,7 @@ export const warn = (message?: any, ...optionalParams: any[]) => {
  * This is when something horribly goes wrong.
  * @param message The message to show.
  */
-export const error = (message?: any, ...optionalParams: any[]) => {
+export const error = (message?: unknown, ...optionalParams: unknown[]): void => {
   log(chalk.red(message), ...optionalParams)
 }
 
@@ -59,18 +56,10 @@ export const error = (message?: any, ...optionalParams: any[]) => {
  * This is for devs only.
  * @param message The message to show.
  */
-export const debug = (message: any, title: string = 'DEBUG') => {
-  log(
-    chalk.magenta(
-      `↓↓↓ --------------------[ ${title} ]-------------------- ↓↓↓`
-    )
-  )
+export const debug = (message: unknown, title = 'DEBUG'): void => {
+  log(chalk.magenta(`↓↓↓ --------------------[ ${title} ]-------------------- ↓↓↓`))
   log(message)
-  log(
-    chalk.magenta(
-      `↑↑↑ --------------------[ ${title} ]-------------------- ↑↑↑`
-    )
-  )
+  log(chalk.magenta(`↑↑↑ --------------------[ ${title} ]-------------------- ↑↑↑`))
 }
 
 /**
@@ -84,11 +73,11 @@ export const pad = (input: string, width: number): string => {
 }
 
 /**
- * Indent message.
- * @param input indent text
+ * Table message.
+ * @param obj indent text
  * @param size indent size
  */
-export const table = (obj: DynamicObject, minCels: number = 20): string => {
+export const table = (obj: Record<string, unknown>, minCels = 20): string => {
   const keys = Object.keys(obj)
   minCels = Math.max(minCels, ...keys.map(k => k.length))
   return keys.map(k => `${pad(k, minCels)} ${obj[k]}`).join('\n')
@@ -99,25 +88,23 @@ export const table = (obj: DynamicObject, minCels: number = 20): string => {
  * @param input indent text
  * @param size indent size
  */
-export const indent = (input: string, size: number = 2): string => {
+export const indent = (input: string, size = 2): string => {
   return redent(input, size)
 }
 
 /**
  * Print a blank line.
  */
-export const newline = () => {
+export const newline = (): void => {
   log('')
 }
 
 /**
  * Prints a divider line
  */
-export const divider = () => {
+export const divider = (): void => {
   log(
-    chalk.gray(
-      '--------------------------------------------------------------------------------'
-    )
+    chalk.gray('--------------------------------------------------------------------------------')
   )
 }
 
@@ -125,7 +112,7 @@ export const divider = () => {
  * Clear console.
  * @param title Default title
  */
-export const clear = (title?: string) => {
+export const clear = (title?: string): void => {
   if (!process.stdout.isTTY) return
   const blank = '\n'.repeat(process.stdout.rows || 30)
   log(blank)
@@ -139,7 +126,7 @@ export const clear = (title?: string) => {
  * @param options The text for the spinner or an ora options.
  * @returns The Ora spinner.
  */
-export const spin = (options: string | object = ''): Ora => {
+export const spin = (options: string | Record<string, unknown> = ''): Ora => {
   return ora(options).start()
 }
 
