@@ -1,44 +1,41 @@
 import { join } from 'path'
 import { readdirSync } from 'fs'
-
-import { Command, Commands } from './types'
+import { Command } from './types'
 
 /**
  * Load dir commands
  * @param dir dir to load
  */
-const loadCommands = (dir: string): Commands => {
+const loadCommands = (dir: string): Record<string, Command> => {
   const commandDir = join(__dirname, dir)
 
   // scanning commands
-  return readdirSync(commandDir).reduce(
-    (cmds, item) => {
-      try {
-        const command: Command = require(join(commandDir, item)).default
-        cmds[command.name] = command
-        if (typeof command.alias === 'string') {
-          cmds[command.alias] = command
-        } else if (command.alias) {
-          command.alias.forEach(a => {
-            cmds[a] = command
-          })
-        }
-      } catch (e) {}
-      return cmds
-    },
-    {} as any
-  )
+  return readdirSync(commandDir).reduce((cmds, item) => {
+    try {
+      // eslint-disable-next-line @typescript-eslint/no-var-requires
+      const command: Command = require(join(commandDir, item)).default
+      cmds[command.name] = command
+      if (typeof command.alias === 'string') {
+        cmds[command.alias] = command
+      } else if (command.alias) {
+        command.alias.forEach(a => {
+          cmds[a] = command
+        })
+      }
+    } catch (e) {}
+    return cmds
+  }, {} as Record<string, Command>)
 }
 
 /**
  * All core commands
  */
-export const coreCommands: Commands = loadCommands('./commands')
+export const coreCommands: Record<string, Command> = loadCommands('./commands')
 
 /**
  * All commands
  */
-export const userCommands: Commands = loadCommands('../commands')
+export const userCommands: Record<string, Command> = loadCommands('../commands')
 
 /**
  * Load command by name
